+++ src/collection.ts
@@ -1,47 +1,9 @@
 import Zotero from 'zotero-lib';
 import fs from 'fs';
-import { ZoteroItem, addItemToCollection } from './addItemToCollection';
+import { addItemToCollection } from './addItemToCollection';
 import { validateMatchfield } from './utils/validateParameters';
-
-type CommanderOptions = {
-  item: string[];
-  itemsfromcollection: string;
-  itemswithtag: string;
-  itemswithouttag: string;
-  itemsfromlibrary: boolean;
-  matchfield?: string[];
-  collection: string[];
-  group: string;
-  test: boolean;
-  name: string;
-  json?: string;
-  ignoretag?: string[];
-  addtag?: string[];
-  recursive?: boolean;
-};
-
-type Collection = {
-  data: {
-    name: string;
-    key: string;
-  };
-  meta?: { numCollections: number; numItems: number };
-  children: Collection[];
-};
-
-type Options = {
-  key: string[];
-  top?: boolean;
-  verbose?: boolean;
-  recursive?: boolean;
-};
-
-type ZoteroCollections = {
-  terms: { term: string; type: string }[];
-  collection: string;
-  collection_name: string;
-  situation: string;
-}[];
+import { CommanderOptions, Collection, Options, ZoteroCollections } from './types/collection';
+import { ZoteroItem } from './types/addItemToCollection';
 
 async function collection(commanderOptions: CommanderOptions) {
   let FinalOutput = '';
@@ -147,7 +109,7 @@ async function collection(commanderOptions: CommanderOptions) {
         collection_name: collectionkey.data.name,
         situation: 'nothing',
       });
-      
+
       if (options.recursive && collectionkey.children?.length > 0) {
         await addChildren(listCollections, collectionkey);
       }
@@ -182,4 +144,4 @@ async function collection(commanderOptions: CommanderOptions) {
   }
 }
 
-export { collection, CommanderOptions };
+export { collection };
